Externalize PR constants into configuration
Does a little extra work to start building toward a fully *composed*
configuration, where all dependency chains, etc. are expressed through
the dependency registration, and the work loop consists of grabbing the
list of repositories, getting the right configured worker based on the
repo kind, and kicking it all off.